pyproject.toml
[project]
name = "pelicantpc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "micawber>=0.5.6",
    "mysql-connector-python>=9.3.0",
    "pelican[markdown]>=4.11.0",
    "typogrify>=2.1.0",
]
